FIX: losing lat/lon precision (10e-7) (#87)
Upstream uses a decimal/integer [i32 representation, scaled to
10e-7](https://github.com/protomaps/PMTiles/blob/main/spec/v3/spec.md#34-position),
just like OSM's native coordinate representation.

We were using f32, which is not precise enough to represent every number
in that range - f64 is. See the test differences in this commit for
examples of the precision loss.

An alternative approach could adopt an i32 backed decimal type. I
actually did something like this for [another crate a while
back](https://github.com/amandasaurus/osmio/pull/4/files#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L39).
In theory this would make for more efficient representation, but as this
is only some fields in the _header_ (as opposed to say, feature
geometries which have *lots* of coordinates), saving a handful of bytes
doesn't seem worth the ceremony.

+++ src/writer.rs
@@ -134,7 +134,7 @@ impl PmTilesWriter {
 
     /// Set the bounds of the tiles
     #[must_use]
-    pub fn bounds(mut self, min_lon: f32, min_lat: f32, max_lon: f32, max_lat: f32) -> Self {
+    pub fn bounds(mut self, min_lon: f64, min_lat: f64, max_lon: f64, max_lat: f64) -> Self {
         self.header.min_latitude = min_lat;
         self.header.min_longitude = min_lon;
         self.header.max_latitude = max_lat;
@@ -151,7 +151,7 @@ impl PmTilesWriter {
 
     /// Set the center position.
     #[must_use]
-    pub fn center(mut self, lon: f32, lat: f32) -> Self {
+    pub fn center(mut self, lon: f64, lat: f64) -> Self {
         self.header.center_latitude = lat;
         self.header.center_longitude = lon;
         self
